requirements.txt
scikit-learn~=1.2.2
Pillow~=9.4.0
Flask~=2.2.3